requirements.txt
# Python 3.9.20
torch==2.5.0+cu118
numpy==1.24.4
matplotlib==3.9.2
opencv-python==4.8.0.76
pyyaml==6.0.2
pandas==2.2.3
seaborn==0.13.2